requirements.txt
streamlit
requests
beautifulsoup4
langchain
langchain-openai
langchain-community
langchain-text-splitters
openai
tiktoken
faiss-cpu
python-dotenv
readability-lxml
streamlit
aiohttp
beautifulsoup4
python-readability
langchain
langchain-community
langchain-text-splitters
sentence-transformers
faiss-cpu
python-dotenv
huggingface-hub